fix: se agregran los dotfiles de sddm en la instalación

+++ .install/displaymanager.sh
@@ -1,29 +1,33 @@
-disman=0
-# Intenta forzar la instalación de sddm
-info_message "Instala sddm"
-yay -S --noconfirm sddm sddm-sugar-candy-git --ask 4
+message "$(figlet 'sddm')"
 
-if [ -f /etc/systemd/system/display-manager.service ]; then
-    sudo rm /etc/systemd/system/display-manager.service
-fi
-sudo systemctl enable sddm.service
+if gum confirm "¿Deseas instalar nuevamente sddm?" ;then
+    disman=0
+    # Intenta forzar la instalación de sddm
+    info_message "Instala sddm"
+    yay -S --noconfirm sddm sddm-sugar-candy-git --ask 4
 
-if [ ! -d /etc/sddm.conf.d/ ]; then
-    sudo mkdir /etc/sddm.conf.d
-    message "Directorio /etc/sddm.conf.d creado."
-fi
+    if [ -f /etc/systemd/system/display-manager.service ]; then
+        sudo rm /etc/systemd/system/display-manager.service
+    fi
+    sudo systemctl enable sddm.service
 
-sudo cp sddm/sddm.conf /etc/sddm.conf.d/
-message "Archivo /etc/sddm.conf.d/sddm.conf actualizado."
+    if [ ! -d /etc/sddm.conf.d/ ]; then
+        sudo mkdir /etc/sddm.conf.d
+        message "Directorio /etc/sddm.conf.d creado."
+    fi
 
-if [ -f /usr/share/sddm/themes/sugar-candy/theme.conf ]; then
+    sudo cp dotfiles/.config/sddm/sddm.conf /etc/sddm.conf.d/
+    message "Archivo /etc/sddm.conf.d/sddm.conf actualizado."
 
-    # Archivo cache para mantener el wallpaper actual
-    sudo cp wallpapers/default.jpg /usr/share/sddm/themes/sugar-candy/Backgrounds/current_wallpaper.jpg
-    echo "El fondo de pantalla por defecto se copió dentro de /usr/share/sddm/themes/sugar-candy/Backgrounds/"
+    if [ -f /usr/share/sddm/themes/sugar-candy/theme.conf ]; then
 
-    sudo cp sddm/theme.conf /usr/share/sddm/themes/sugar-candy/
-    sudo sed -i 's/CURRENTWALLPAPER/'"current_wallpaper.jpg"'/' /usr/share/sddm/themes/sugar-candy/theme.conf
-    echo "Archivo theme.conf actualizado en /usr/share/sddm/themes/sugar-candy/"
+        # Archivo cache para mantener el wallpaper actual
+        sudo cp wallpapers/default.jpg /usr/share/sddm/themes/sugar-candy/Backgrounds/current_wallpaper.jpg
+        echo "El fondo de pantalla por defecto se copió dentro de /usr/share/sddm/themes/sugar-candy/Backgrounds/"
 
+        sudo cp dotfiles/.config/sddm/theme.conf /usr/share/sddm/themes/sugar-candy/
+        sudo sed -i 's/CURRENTWALLPAPER/'"current_wallpaper.jpg"'/' /usr/share/sddm/themes/sugar-candy/theme.conf
+        echo "Archivo theme.conf actualizado en /usr/share/sddm/themes/sugar-candy/"
+
+    fi
 fi
